Simplify Dashboard component
